+++ R/dump_fontdb.R
@@ -5,7 +5,6 @@
 #'
 #' @return A `tibble()` containing these columns:
 #' \describe{
-#'   \item{source}{The source file of the font face.}
 #'   \item{index}{The index of the font face within the source.}
 #'   \item{family}{The font family of the face.}
 #'   \item{weight}{The weight of the face.}